Switched to float64

+++ src/train.py
@@ -35,7 +35,7 @@ def train_model(args: argparse.Namespace) -> None:
         args.hidden_size,
         args.num_layers,
         args.num_attn_heads
-    ).to(device)
+    ).to(torch.float64).to(device)
     train_dataset, validation_dataset, test_dataset = load_data(args.data_path, args.seed, args.frac_train,
                                                                 args.frac_validation, args.frac_test,
                                                                 args.use_small_dataset)
@@ -103,7 +103,7 @@ def run_training_epoch(progress_bar: tqdm, optimizer: optim.Optimizer, model: nn
 
     for batch_data in progress_bar:
         node_features, edge_features, adjacency_matrix, pchembl_score = [
-            x.to(torch.float32).to(device) for x in batch_data
+            x.to(torch.float64).to(device) for x in batch_data
         ]
 
         preds = model(node_features, edge_features, adjacency_matrix).squeeze(-1)
@@ -135,7 +135,7 @@ def get_validation_metrics(validation_loader: DataLoader, model: nn.Module, loss
 
     for batch in validation_loader:
         node_features, edge_features, adjacency_matrix, pchembl_scores = [
-            x.to(torch.float32).to(device) for x in batch
+            x.to(torch.float64).to(device) for x in batch
         ]
         preds = model(node_features, edge_features, adjacency_matrix).squeeze(-1)
         loss = loss_func(preds, pchembl_scores).item()
